AgroScope/backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()

API_KEY = os.getenv("DEEPSEEK_API_KEY", "")
AI_MODEL = os.getenv("AI_MODEL", "deepseek-chat")
API_URL = "https://openrouter.ai/api/v1/chat/completions"

SYSTEM_PROMPT = """
Ты — эксперт-агроаналитик AgroScope.
Анализируешь фото поля, определяешь количество упавших зерен и рассчитываешь потери урожая.
Используй формулу:
Потери (кг/га) = (N × 0.000045 × 10000) / S
где:
N — количество найденных зерен,
S — площадь изображения в м² (0.25 по умолчанию).
Площадь поля в гектарах = user_input_field_area.
Ответ верни строго в JSON:
{
  "grain_count": <int>,
  "losses_per_ha": <float>,
  "season_loss": <float>
}
"""
