AgroScope/backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pillow
